requirements.txt
dash_html_components==1.1.2
scipy==1.4.1
plotly==4.14.3
pandas==1.1.2
numpy==1.18.1
dash==1.19.0
matplotlib==3.3.2
scikit_image==0.16.2
dash_core_components==1.15.0
opencv_python==4.5.1.48
Pillow==8.1.0
skimage==0.16.2
